+++ routes/groups/groups.js
@@ -1,6 +1,6 @@
-const controller = require("../controllers/controller");
+const controller = require("../../controllers/groupsController");
 const express = require("express");
 const router = express.Router();
-const middleware = require("../middleware/middleware");
+const middleware = require("../../middleware/middleware");
 router.get("/", controller.getgroups);
 module.exports = router;
