+++ routes/groups/joingroup.js
@@ -1,0 +1,6 @@
+const controller = require("../../controllers/groupsController");
+const express = require("express");
+const router = express.Router();
+const middleware = require("../../middleware/middleware");
+router.post("/", middleware.verifyToken, controller.joingroup);
+module.exports = router;
