+++ routes/invitations/acceptinvitation.js
@@ -1,4 +1,4 @@
-const controller = require("../controllers/controller");
+const controller = require("../controllers/invitationsController");
 const express = require("express");
 const router = express.Router();
 router.post("/", controller.acceptinvit);
