--- routes/sendmessage.js
@@ -1,6 +1,0 @@
-const controller = require("../controllers/controller");
-const express = require("express");
-const router = express.Router();
-const middleware = require("../middleware/middleware");
-router.post("/", middleware.verifyToken, controller.addmessage);
-module.exports = router;
