+++ routes/user/updateuser.js
@@ -1,6 +1,6 @@
-const controller = require("../controllers/controller");
+const controller = require("../controllers/userController");
 const express = require("express");
 const router = express.Router();
 const middleware = require("../middleware/middleware");
-router.post("/", middleware.verifyToken, controller.joingroup);
+router.put("/", middleware.verifyToken, controller.updateuserinfo);
 module.exports = router;
