+++ server.js
@@ -7,25 +7,25 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 server.listen(3001);
 const cookieParser = require("cookie-parser");
-const group = require("./routes/groups");
-const users = require("./routes/users");
-const user_messages = require("./routes/user_messages");
-const deletemessage = require("./routes/deletemessage");
-const addmessage = require("./routes/sendmessage");
+const group = require("./routes/groups/groups");
+const users = require("./routes/user/users");
+const user_messages = require("./routes/user/user_messages");
+const deletemessage = require("./routes/user/deletemessage");
+const addmessage = require("./routes/user/sendmessage");
 const login = require("./routes/login");
-const user = require("./routes/connecteduser");
+const user = require("./routes/user/connecteduser");
 const register = require("./routes/register");
-const addgp = require("./routes/addgroup");
-const groups = require("./routes/groups");
-const joingroup = require("./routes/joingroup");
-const addmessagetogroup = require("./routes/addmessagetogroup");
-const groupsmsg = require("./routes/groupsmessages");
-const updateuserprofile = require("./routes/updateuser");
-const editpass = require("./routes/editpassword");
-const addfriend = require("./routes/addfriend");
-const acceptinvit = require("./routes/acceptinvitation");
-const removefriend = require("./routes/deletefriend");
-const acceptinvit2=require('./routes/acceptinvit2')
+const addgp = require("./routes/groups/addgroup");
+const groups = require("./routes/groups/groups");
+const joingroup = require("./routes/groups/joingroup");
+const addmessagetogroup = require("./routes/groups/addmessagetogroup");
+const groupsmsg = require("./routes/groups/groupsmessages");
+const updateuserprofile = require("./routes/user/updateuser");
+const editpass = require("./routes/user/editpassword");
+const addfriend = require("./routes/invitations/addfriend");
+const acceptinvit = require("./routes/invitations/acceptinvitation");
+const removefriend = require("./routes/invitations/deletefriend");
+const acceptinvit2 = require("./routes/invitations/acceptinvit2");
 app.use("/list", group);
 app.use("/users", users);
 app.use("/messages", user_messages);
@@ -44,5 +44,5 @@ app.use("/updatepassword", editpass);
 app.use("/sentinvit", addfriend);
 app.use("/acceptinvit", acceptinvit);
 app.use("/removefriend", removefriend);
-app.use('/acceptinvit2',acceptinvit2)
+app.use("/acceptinvit2", acceptinvit2);
 app.use(cookieParser);
